Add support to update content attempt
